test: move profile command errors in process
